alter group paths centrally
This will work for any output including json, etc. The `link_alter()` method was only being applied to html output. As far as I can tell with my testing (which might be lacking) this appears to still work in all the same cases that `link_alter()` was implementing. By the time a node/x reaches the `processOutbound()` method in this module, other `processOutbound()` implementations would have already converted it to the /group/blah path, so our `processOutbound()` method will work as expected.

This should undergo heavy testing before it is merged, and probably also deserves unit tests